(PC-18749)[API] refactor: rename file to reminder_emails
