+++ api/src/pcapi/core/subscription/commands.py
@@ -1,6 +1,6 @@
 import logging
 
-from pcapi.core.mails.transactional.users.ubble.ubble_ko_reminder import send_reminder_emails
+from pcapi.core.mails.transactional.users.ubble import reminder_emails
 from pcapi.scheduled_tasks.decorators import log_cron_with_transaction
 from pcapi.utils.blueprint import Blueprint
 
@@ -12,4 +12,4 @@ logger = logging.getLogger(__name__)
 @blueprint.cli.command("send_beneficiary_subscription_reminders")
 @log_cron_with_transaction
 def send_beneficiary_subscription_reminders() -> None:
-    send_reminder_emails()
+    reminder_emails.send_reminder_emails()
